+++ MASFoundation/Classes/MAS.m
@@ -33,6 +33,51 @@
 }
 
 
++ (void)setDeviceRegistrationType:(MASDeviceRegistrationType)registrationType
+{
+    switch (registrationType) {
+            
+        case MASDeviceRegistrationTypeClientCredentials:
+            [MASModelService setGrantFlow:MASGrantFlowClientCredentials];
+            break;
+            
+        case MASDeviceRegistrationTypeUserCredentials:
+            [MASModelService setGrantFlow:MASGrantFlowPassword];
+            break;
+            
+        case MASDeviceRegistrationTypeCount:
+            [MASModelService setGrantFlow:MASGrantFlowCount];
+            break;
+            
+        default:
+            [MASModelService setGrantFlow:MASGrantFlowUnknown];
+            break;
+    }
+}
+
+
++ (MASDeviceRegistrationType)deviceRegistrationType
+{
+    switch ([MASModelService grantFlow]) {
+        case MASGrantFlowClientCredentials:
+            return MASDeviceRegistrationTypeClientCredentials;
+            break;
+        
+        case MASGrantFlowPassword:
+            return MASDeviceRegistrationTypeUserCredentials;
+            break;
+            
+        case MASGrantFlowCount:
+            return MASDeviceRegistrationTypeCount;
+            break;
+            
+        default:
+            return MASDeviceRegistrationTypeUnknown;
+            break;
+    }
+}
+
+
 + (void)setGrantFlow:(MASGrantFlow)grantFlow
 {
     [MASModelService setGrantFlow:grantFlow];