+++ MASFoundation/Classes/MASConstants.h
@@ -86,6 +86,33 @@ typedef void (^MASUserLoginWithUserCredentialsBlock)(MASBasicCredentialsBlock ba
 # pragma mark - MAS Constants
 
 /**
+ * The enumerated MASRegistrationTypes.
+ */
+typedef NS_ENUM(NSInteger, MASDeviceRegistrationType)
+{
+    /**
+     * Unknown encoding type.
+     */
+    MASDeviceRegistrationTypeUnknown = -1,
+    
+    /**
+     * The client credentials registration type.
+     */
+    MASDeviceRegistrationTypeClientCredentials,
+    
+    /**
+     * The user credentials registration type.
+     */
+    MASDeviceRegistrationTypeUserCredentials,
+    
+    /**
+     * The total number of supported types.
+     */
+    MASDeviceRegistrationTypeCount
+} DEPRECATED_MSG_ATTRIBUTE("Use MASGrantFlow instead.");
+
+
+/**
  * The enumerated MASGrantFlow.
  */
 typedef NS_ENUM(NSInteger, MASGrantFlow)