+++ MASFoundation/Classes/models/MASDevice.m
@@ -121,6 +121,25 @@ static id<MASSessionSharingDelegate> _SessionSharingDelegate_;
 }
 
 
+- (void)logOutDeviceAndClearLocal:(BOOL)clearLocal completion:(MASCompletionErrorBlock)completion
+{
+    //
+    // If the user is not authenticated, return an error
+    //
+    if (![MASUser currentUser])
+    {
+        if (completion)
+        {
+            completion(NO, [NSError errorUserDoesNotExist]);
+        }
+    }
+    else {
+        
+        [[MASUser currentUser] logoutWithCompletion:completion];
+    }
+}
+
+
 - (void)resetLocallyWithCompletion:(MASCompletionErrorBlock)completion
 {
     //